requirements.txt
Flask
pandas
line-bot-sdk
langchain
google-generativeai
langchain-google-genai
